Chore: Prep for scarf.sh
Scarf does their usage analysis by accepting requests and returning redirects from Scarf to an actual handler/source. This allows them to inspect requests without having to serve as a host or full proxy.

The alternative to this approach of tolerating Scarf's redirects would be to do IP logging ourselves and ingest to Scarf on the side, which exposes us to warehousing PII.