Feat: Allow users to salt values (#12)
This came up in the original rules authors SIG call where we reviewed
this system.

Some users expressed concern that sha1 is a computationally broken
(tractable under brute force) hash, and that as such correlation IDs are
vulnerable to rainbow table or other attack which could lead to the
exfiltration of project names or other values considered sensitive by
some organizations.

Providing support for user-defined salting allows for users to protect
potentially sensitive values without disabling telemetry features
entirely and document how to do so as part of the `README`.

As an alternative to this we considered but rejected allowing users to
specify override values for specific tracking features. Allowing for
consistent salting of IDs addresses the same concern without
complicating setup or implementation.

### Changes are visible to end-users: yes

- Suggested release notes appear below: yes

Added support for user-defined salting of correlation IDs.

### Test plan

- Manual testing